chore(geo): more v6 migration, turn on ts strict
