+++ packages/geo/tsconfig.build.json
@@ -1,5 +1,8 @@
 {
 	"extends": "../tsconfig.base.json",
-	"compilerOptions": {},
+	"compilerOptions": {
+		"strict": true,
+		"typeRoots": ["./node_modules/@types", "../../node_modules/@types"]
+	},
 	"include": ["lib*/**/*.ts", "src"]
 }
