+++ packages/geo/tsconfig.json
@@ -1,0 +1,9 @@
+{
+	"extends": "../tsconfig.base.json",
+	"compilerOptions": {
+		"strict": true,
+		"typeRoots": ["./node_modules/@types", "../../node_modules/@types"],
+		"types": ["node", "jest"]
+	},
+	"include": ["./src"]
+}
